All the project
This is the entire content of my project related to building a similar image query system from data available on the system. Note that it is still quite simple and it only performs brute-force feature matching, not yet improved with algorithms such as Kmeans clustering, Inverse Index Retrieval,... On github, I can only push the Histogram features that I computed because the ORB or SIFT is kinda large. :(
The dataset that I used is TmBUD-main. This is the link: https://github.com/CipiOrhei/TMBuD